Take a first pass at the service manager
This commit has quite a few little bits and pieces to be honest, however
the bulk of the work was in setting up the first version of the service
manager.

You can find the work in runtime/service. It's quite simple really, it
is just a small utility for managing a subprocess, and a gRPC client
for communicating with that subprocess over a unix domain socket.

Other bits include: a logging interface, a few bits of the HTTP handler,
and wiring it up to the foldrt binary.