Add router and start on go sdk
This commit adds quite a lot. It adds the first version of the router
which is working alright. It also changes the interface and organisation
of the runtime package a fair bit, it's a bit better now, it will do
for now.

Additionally, this starts on adding in a go sdk. I was planning to
do a TS one first, and I will, but the go one will be very simple so
I figured I would knock it out to make it easier to test out some of
the runtime code in the meantime.