Tidy up router tests
The router tests were quite embarassing and turned out to be flaky
too. This commit just sorts that out by marshaling the response
bodies to and from json.

It also brings in the go-cmp library for diffing the response bodies,
which is very nice actually. I will use that more.