Implement AWS Lambda handler
This implements and tests the AWS Lambda Handler. I've yet to test
it out 'for real' so changes are probably due. However I've got fairly
high confidence in it as the prototype worked fine and was very
similar.

Additionally, the PR tidies up the runtime.go file, adds the handler
branch into the foldrt main, and moves some json handling code
to a testutils package as it's quite handy.