Implement v1 go sdk
I was meant to implement the sdk in TypeScript first, however it
was too easy to write it in go given that I had the gRPC code figured
out etc. I wanted a way to test out the runtime properly and so just
sticking this basic go sdk in seemed the easiest way.

I will still implement it in TypeScript though, as that is what will
appeal to my target user base. Shouldn't take long now that I've
figured out how it will all work.

The commit also adds an example for go to the examples directory.