Handle SIGTERM and SIGINT in the runtime
This commit ensures these signals are handled appropriately in the
runtime and tidy up resources like the socket. It still doesn't pass
the signal through to the subprocess, we can do that later.

It also tidies up a couple of small things around the code base and
adds another example for the go sdk which demonstrates a little
more of the API.