Add an SDK for node js
This PR basically reimplements the go SDK in TypeScript.
It is very basic but offers the same features as the go one
and works nicely with the runtime at the moment.

There will be a fair bit to add I'm sure to make it really robust.
Additionally, the API is a bit poor, I will study express for
some ideas of how to make it a bit better. It would be nice
to parse bodies based on type parameters in handlers.

Spent quite a while trying to get the communication working
between the runtime, turns out various gRPC implementations
handle things a little differently for unix sockets. May
have to move to using localhost at some point.

There are a few other changes which were all in the spirit of
making the node sdk work, and I've also added a working example
for node.