Start on foldctl
This commit has done a lot, got a bit carried away in a branch.
The gist of it is, however, though I have started on the command line
application and implemented the following commands:

- new
- init
- build
- version

There's a lot to be improved on but I'm happy enough that the pattern
is about right to commit and keep going from here. I'm starting
to build up quite a range of utilities that are making the commands
easier to write and a bit less verbose. It also helps ensure
somewhat consistent output from the commands. Not sure what to do
about help text and error messages. That gets very messy and obscures
the logic. I've got a few standard error messages, could take that
the whole way I guess and just have one file which stores loads of
strings?

I also set up a templates repository which the new command uses.