Add foldgw and update Makefile
This commit mostly adds the first bit of the fold gateway cmd
and makes some changes to the Makefile to reflect how I'm now
actually building stuff.

There are a number of small changes that are preparing this for
the next step though, as well as just tidying a few things up in
foldctl.