Add container and network support to docker client
This isn't complete yet and hasn't been tested, but it's roughly
all the functionality I think I'll need. Will need to spend time
fiddling with all the options no doubt.

As well as getting the rough docker client functionality in place,
I've also added a thin Container/Network abstraction which I will
be able to expose to the cli tool. Should make it quite easy in there
to start and stop containers, join them to networks, etc.

Not sure how to test all of this, the docker client API is massive,
and it's not clear to me if mocking it and making assertions about
what is being called would actually give me any meaningful assurance.

Perhaps mocking the actual REST endpoint, capturing output and snapshotting
it could be useful?