Add docker interface and up and down commands
This PR made an interface for the bits of the docker client I'm
actually using and then introduced gomock to mock it out. It's
quite nice actually, somewhat heavy weight with the code generation
though so will probs use it sparingly.

Once that was done and the container package was working well
it was pretty easy to to implement the up and down commands.
So I did that too.