Refactor of project and container
Before this I was ending up with a load of code in the commands
for wiring up the project and container code, which was rubbish,
and hard to test.

I thought a better way to do it might be to treat the project as
the main interface for manipulating a fold project, and have that
encapsulate all the logic about how to build, run, etc fold services.

It's much nicer now I think, and a fair bit simpler too. Now time
to get on with some real work...