Refactor project package and improve error reporting
Couldn't resist and kept on refactoring... An issue with moving
logic out of the commands is that all of the old error reporting
and status updates were removed. Not good, users need feedback!

My solution to this was to hijack the logger for the normal use
case and give it a logformat that just displays messages. Additionally,
I've started moving to using error values directly for printing out
error messages.

I quite like doing it this way. It means that progress updates and
errors can be defined naturally along with the code that produces them.
However, because it is all done through the logger and error values,
the library code doesn't actually know that it is producing output for a
user of the cli tool.

I.e. the same logging and error messages are just as meaningful if the
code is run in a headless way on, for example, a build server. All you
need to do is inject a more tradtionally configured logger and you get
meaningful info level logging.

I wonder if there is a downside to doing it like this... seems good to
me right now though.