Add gateway to up and down commands
This PR introduced everythign for stopping and starting the
gateway service along with the network. This involved a little
bit of new code in the project, and a tidy up of the image
building and pulling code in the container package.

I also moved the SemVer struct out into its own little version package
at the top as it's being used in three places now. At the same time,
I made a ctl/gateway package to house both the gateway server code
and info about the version , etc, all in one place.