Configure containers
This commit has made a few changes to configure containers so
that they are actually fit for purpose. They now:
  - Have network aliases for the gateway to use
  - Mount volumes
  - Map ports to the host machine (currently only allowed for the
    gateway)
  - Autoremove
  - Automatically join the fold network on creation

After learning a little more about the docker API, I was able to
simplify the interface used by the project a bit. Namely, I removed
the ability to explicitly add a container to a network, and the ability
to remove a container.