Implement gateway proxy
This commit adds in a functioning gateway proxy. I did it in a
very quick and dirty way using the reverse proxy from the
standard library. It's inefficient etc but for the local dev
use case it is fine and it is very little code to look after.

There were a few other little fixes required around the place
to make it work, mostly with Dockerfiles.