Implement hot reload, write README
This commit gets the basic hot reloading working for the runtime.
I probably shouldn't have bothered with that at this stage to be
honest. It's not perfect but it's quite good and works surprisingly
well.

I've also written up a fair bit in the README so that there is some
indication of how to use the thing.

Additionally, I modified the Makefile to build and tag all the relevant
images so that I can do the first 'release'.

Outside of that, this commit fixed a few bugs and simplified a few
things. Bit of a mix as always really, need to work on that.