Control how project is  written to disk
There was previoulsy an issue where the Port was being written to
to the fold config file. This was not intentional and was happening
purely because the Port propery was public on the struct.

To solve it I have made an 'on disk' struct which represents what
gets stored to the disk and the config code marshalls/unmarshalls
to/from it. This makes it resistant to future changes in the internal
Project or Service structs.