Create host dir for mount if it doesn't exist
When you run foldctl up with a mount specified it will now create
the directory on the host if it doesn't already exist. This is
what the docker cli does see it seems like pretty reasonable
behaviour.

If you don't do it then it leads to poor UX with the up command.
I.e., you add the mount for a generated directory (e.g. build out)
and then you run up and the container fails to start as docker can't
bind something that doesnt' exist.