Add --version flag to print version info
This doesn't remove the version command, it just adds the --version
flag as well. They both behave in exactly the same way.

While many of the CLI tools written in go don't support --version,
it's very common outside of it so this makes the tool more intuitive
to some users.