Consolidate init and new into one command
The init and new commands were a bit confusing before. This commit
consolidates them under one command, called new. The idea is that
it is an umbrella command for creating fold resources in general.

As such, it has two subcommands, project and service, which do what
you would expect.

Additionally, I have made the new service command work with a prompt
in the same way as the new project command. It's much better like
that and makes the validation logic easy.