Create build script and simplify tar structure
The Makefile was getting a bit grim for me, and it will need to
do more stuff in the near future. As such, I've moved the build
logic into a shell script, using the Makefile as a runner for the
script. It should be easier to look after now but we'll see how it
goes.