Complete MVP Node SDK
The vast majority of this commit is work on the node sdk in response
to feedback from a couple of early testers. It is quite good now, and
sticks to the express interface quite well, it even supports quite a few
3rd party express middlewares.

In doing this work, I worked through a number of issues with the HTTP
aspects of the protobuf manifest, so that is all much more fully
featured now and the changes are reflected in the runtime code.