Reimplement runtime as FSM
This commit completely replaces most of the runtime with an
implementaiton based on a simple finite state machine. The old one was
poorly done and it was difficult to manage the relationship between the
state of the child process and the expected behaviour of the runtime.

Carving out the state of the runtime into an FSM and putting all of the
dependencies behind interfaces has made the whole thing much easier to
reason about and it's possible to unit test it quite effectively now.

I've also added in  a few integration tests which exercise the whole
runtime and its subprocess in a very realistic way. I will keep adding
to those over time. Ideally I want to extend them to incorporate the
whole workflow but these are a good start.

Additionally, the commit fixes a bug with the template cloning code, it
was trying to be too clever so I have just made it delete and clone the
templates every time.