feat(ctl): run up in foreground by default
This commit runs up in the foreground by default. It also contains a
whole load of other stuff as I completely refactored the CLI app and
bits of the ctl package in order to do it. Still lots of room for
improvement of course but it's definitely better.

Part of this was adding a lot to the 'output' system which was
previously just a few dubious functions. There is now a semi decent
utility object for writing output to the terminal. It handles
multiplexing streams of text and formatting and that kind of thing.

I've also moved all the tests over the testify/mockery. It definitely
makes the tests quicker and easier to write than just the stdlib so I
think it's worth it.